pyproject.toml
[build-system]
requires = ["setuptools>=68,<81", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wandavoice"
version = "0.2.1"
description = "VOX Voice System (Gateway + Orb + UI)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "setuptools<81",
  "click>=8.1",
  "numpy>=1.24",
  "sounddevice>=0.4.6",
  "soundfile>=0.12.0",
  "webrtcvad>=2.0.10",
  "faster-whisper>=1.0.3",
  "PyYAML>=6.0",
  "pyperclip>=1.8.2",
  "pynput>=1.7.6",
  "evdev>=1.6.0",
  "azure-cognitiveservices-speech>=1.34.0",
  "requests>=2.31",
  "torch>=2.0",
  "torchaudio>=2.0",
  "librosa>=0.10",
  "websockets>=12.0",
  "edge-tts>=6.0",
]

[project.scripts]
vox = "wandavoice.main:main"
voice = "wandavoice.main:main"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["wandavoice"]
